requirements.txt
telethon
asyncio
input
smsactivateru
names
python-dotenv
pytgcalls
request